Restructure
Restructured code so that is supports each user having his own model

+++ src/main.py
@@ -1,3 +1,4 @@
+import platform
 from model import Model
 from flask import Flask, request, jsonify
 import os
@@ -8,38 +9,40 @@ app = Flask(__name__)
 # Current path "http_://127.0.0.1:5000/face-recognition/authenticate"
 @app.route('/face-recognition/authenticate', methods=['POST'])
 def authenticate():
-    # Path might need to be changed
-    model_path = "Project-Face-ID/models/model.pt"
-    temp_image_name = "temp_img.png"
-
-    #model = Model()
-    #model.load(model_path)
-
-    if 'image' not in request.files:
-        return jsonify({"error": "No image part in the request"}), 400
-
-    img = request.files['image']
-
-    if img.filename == '':
-        return jsonify({"error": "No image path found"}), 400
-    
     if 'user' not in request.form:
         return jsonify({"error": "No user part in the request"}), 400
-    
     user = request.form['user']
-    
-    img.save(temp_image_name)
-    abs_img_path = os.path.abspath(temp_image_name)
 
-    #result = model.eval(abs_img_path)
-    os.remove(abs_img_path)
+    model_path = f"Project-Face-ID/models/{user}.pt"
+    if not os.path.exists(model_path):
+        # return jsonify({"error": f"Model file {model_path} does not exist"}), 400
+        # Code to train new model
+        pass
+    else:
+        classes = 4
+        model = Model(classes)
+        model.load(model_path)
 
-    #if result == user:
-    #    return jsonify({"result": 1})
-    #else:
-    #    return jsonify({"result": 0})
+        if 'image' not in request.files:
+            return jsonify({"error": "No image part in the request"}), 400
+        img = request.files['image']
 
-    return jsonify({"result": "susy bakka"})
+        if img.filename == '':
+            return jsonify({"error": "No image path found"}), 400
+
+        temp_image_name = "temp_img.png"
+        img.save(temp_image_name)
+        abs_img_path = os.path.abspath(temp_image_name)
+
+        choice, threshold, probs = model.eval(abs_img_path)
+        os.remove(abs_img_path)
+
+        for i, p in enumerate(probs):
+            if i == choice:
+                if p >= threshold:
+                    return jsonify({"result": 1}) # Image class was found
+        
+        return jsonify({"result": 0}) # Image class was not found
 
 if __name__ == '__main__':
     # Add debug=True for debugging
